Added build failure check

+++ run.sh
@@ -13,7 +13,11 @@ fi
 
 SCRIPT_DIR=`dirname $(realpath -m "$0")`
 CONTAINER_NAME="vasiniyo-chat-bot"
-echo "Building $CONTAINER_NAME..."     && docker build -t "$CONTAINER_NAME" "$SCRIPT_DIR" &>/dev/null
+echo "Building $CONTAINER_NAME..." && docker build -t "$CONTAINER_NAME" "$SCRIPT_DIR"
+if [ $? -ne 0 ]; then
+    echo "Failed to build container $CONTAINER_NAME"
+    exit 1
+fi
 echo "Stopping old $CONTAINER_NAME..." && docker stop "$CONTAINER_NAME" 2>/dev/null
 echo "Removing old $CONTAINER_NAME..." && docker rm "$CONTAINER_NAME" 2>/dev/null
 docker run --rm\
